commiting most recent pre-svn oscpack version. Main change is new networking classes with new unified UdpSocket class, support for listening to multiple sockets. New IpEndpointName class for refering to endpoints by name, ip address etc.

+++ oscpack/trunk/osc/OscPrintReceivedElements.cpp
@@ -115,7 +115,7 @@ std::ostream& operator<<( std::ostream & os,
 
                 // strip trailing newline from string returned by ctime
                 const char *timeString = std::ctime( &t );
-                int len = strlen( timeString );
+                size_t len = strlen( timeString );
                 char *s = new char[ len + 1 ];
                 strcpy( s, timeString );
                 if( len )
